Run Increment Length encoding
* Implement compressor and decompressor in gbagfx tool
* Implement decompressor in gba code

This is a simple compression algorithm that is designed around the attributes
that appear often in tilemaps, particularly the tilemaps for cinematic images

For example, these creates a smaller-than-lz compression on most of the intro
tilemaps, and compresses `device_outline_map.bin` to 52 bytes (compared to
lz's 272 bytes).

`RilXDecompressVram` is the equivalent of `LZ77UnCompVram`. Most graphics use
a wrapper instead of `LZ77UnCompVram` directly, and this commit does not
include a direct equivalent of `DecompressAndCopyTileDataToVram` or other
similar wrappers