Builds secret base decoration tilesets and metatiles dynamically
BG decorations (i.e. not dolls or cushions) no longer have tiles in the
`secret_base` primary and are not represented as a list of metatile ids.
Instead, a decoration is represented as a independent tileset and a metatile
list. When in a secret base, the tilesets and metatiles of the decorations
in the current secret base will be combined to form the tileset and metatiles
of the secret base.

The benefit of this refactor is that the variety of decorations is no longer
limited by tiles per tileset and metatiles per tileset.
The tradeoff is that the number of decorations active at once is limited by
those factors instead - however the vanilla 16 decoration limit still
comfortably fits in a tileset.