Combine middle and top layers when drawing decoration sprite
Before, if a triple-layer decoration metatile quadrant had both a middle layer
and a top layer; when drawing the sprite for the place decoration action,
the top layer tiles would fully overwrite the middle layer.
Now, only the non-transparent pixels in the top layer are drawn to the sprite,
allowing the middle layer to show through.